Refactor with Tenants (#13)
* refactor for tenants

---------

+++ terraform/hub/destroy.sh
@@ -6,28 +6,19 @@ SCRIPTDIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 ROOTDIR="$(cd ${SCRIPTDIR}/../..; pwd )"
 [[ -n "${DEBUG:-}" ]] && set -x
 
-
-pushd () {
-    command pushd "$@" > /dev/null
-}
-
-popd () {
-    command popd "$@" > /dev/null
-}
-
-pushd ${SCRIPTDIR}
-
 # Delete the Ingress/SVC before removing the addons
 TMPFILE=$(mktemp)
-terraform output -raw configure_kubectl > "$TMPFILE"
-source "$TMPFILE"
+terraform -chdir=$SCRIPTDIR output -raw configure_kubectl > "$TMPFILE"
+# check if TMPFILE contains the string "No outputs found"
+if [[ ! $(cat $TMPFILE) == *"No outputs found"* ]]; then
+  echo "No outputs found, skipping kubectl delete"
+  source "$TMPFILE"
+  kubectl delete svc -n argocd argo-cd-argocd-server
+fi
 
-kubectl delete svc -n argocd argo-cd-argocd-server
 
-terraform destroy -target="module.gitops_bridge_bootstrap" -auto-approve
-terraform destroy -target="module.eks_blueprints_addons" -auto-approve
-terraform destroy -target="module.eks" -auto-approve
-terraform destroy -target="module.vpc" -auto-approve
-terraform destroy -auto-approve
-
-popd
+terraform -chdir=$SCRIPTDIR destroy -target="module.gitops_bridge_bootstrap" -auto-approve
+terraform -chdir=$SCRIPTDIR destroy -target="module.eks_blueprints_addons" -auto-approve
+terraform -chdir=$SCRIPTDIR destroy -target="module.eks" -auto-approve
+terraform -chdir=$SCRIPTDIR destroy -target="module.vpc" -auto-approve
+terraform -chdir=$SCRIPTDIR destroy -auto-approve
